# bruno/cms-api-openapi.yaml
openapi: 3.0.3
info:
  title: Complaint Management System API
  version: "1.0.0"
  description: Bruno-importable full API collection for complaint-management-api.
servers:
  - url: https://localhost:7000
    description: Local HTTPS
  - url: http://localhost:5000
    description: Local HTTP
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Login or client email check
      description: If `password` is empty/blank, this checks whether a client email exists. Otherwise performs login.
      operationId: authLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequestDto'
            examples:
              login:
                value:
                  emailOrUsername: admin@cms.local
                  password: Admin@123
              emailCheck:
                value:
                  emailOrUsername: client@cms.local
                  password: ""
      responses:
        '200': { description: Success }
        '400': { description: Bad request }
        '401': { description: Unauthorized }
  /api/v1/auth/client/register:
    post:
      tags: [Auth]
      security: []
      summary: Register client account
      operationId: registerClient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientRegisterRequestDto'
      responses:
        '201': { description: Created }
        '400': { description: Bad request }
...
  /api/v1/complaints:
    get:
      tags: [Complaints]
      summary: Search complaints
      operationId: searchComplaints
      parameters:
        - name: statusId
          in: query
          schema: { type: integer, format: int64 }
        - name: categoryId
          in: query
          schema: { type: integer, format: int64 }
        - name: channelId
          in: query
          schema: { type: integer, format: int64 }
        - name: priority
          in: query
          schema: { type: string }
        - name: assignedToUserId
          in: query
          schema: { type: integer, format: int64 }
        - name: q
          in: query
          schema: { type: string }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200': { description: OK }
    post:
      tags: [Complaints]
      summary: Create complaint
      operationId: createComplaint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateComplaintRequest'
      responses:
        '201': { description: Created }
  /api/v1/complaints/{id}:
    get:
      tags: [Complaints]
      summary: Get complaint by id
      operationId: getComplaintById
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': { description: OK }
        '404': { description: Not found }
    delete:
      tags: [Complaints]
      summary: Delete complaint
      operationId: deleteComplaint
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': { description: Deleted }
  /api/v1/complaints/{id}/history:
    get:
      tags: [Complaints]
      summary: Get complaint history
      operationId: getComplaintHistory
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': { description: OK }
  /api/v1/complaints/{id}/status:
    patch:
      tags: [Complaints]
      summary: Update complaint status
      operationId: updateComplaintStatus
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateComplaintStatusRequest'
      responses:
        '200': { description: Updated }
  /api/v1/complaints/{id}/assign:
    post:
      tags: [Complaints]
      summary: Assign complaint
      operationId: assignComplaint
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignComplaintRequest'
      responses:
        '200': { description: Assigned }
  /api/v1/complaints/{id}/escalate:
    post:
      tags: [Complaints]
      summary: Escalate complaint
      operationId: escalateComplaint
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EscalateComplaintRequest'
      responses:
        '200': { description: Escalated }
  /api/v1/complaints/{id}/resolve:
    post:
      tags: [Complaints]
      summary: Resolve complaint
      operationId: resolveComplaint
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveComplaintRequest'
      responses:
        '200': { description: Resolved }
  /api/v1/complaints/{id}/close:
    post:
      tags: [Complaints]
      summary: Close complaint
      operationId: closeComplaint
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': { description: Closed }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        default: 20
  schemas:
    LoginRequestDto:
      type: object
      properties:
        emailOrUsername:
          type: string
        password:
          type: string
      required: [emailOrUsername]
    ClientRegisterRequestDto:
      type: object
      properties:
        name: { type: string }
        email: { type: string, format: email }
        phoneNumber: { type: string }
        password: { type: string }
        confirmPassword: { type: string }
      required: [name, email, phoneNumber, password, confirmPassword]
...
    CreateComplaintRequest:
      type: object
      properties:
        clientId: { type: integer, format: int64, nullable: true }
        clientName: { type: string, nullable: true }
        clientEmail: { type: string, format: email, nullable: true }
        clientMobile: { type: string, nullable: true }
        complaintChannelId: { type: integer, format: int64 }
        complaintCategoryId: { type: integer, format: int64 }
        subCategoryId: { type: integer, format: int64, nullable: true }
        subject: { type: string }
        description: { type: string }
        priority: { type: string, example: Medium }
      required: [complaintChannelId, complaintCategoryId, subject, description, priority]
    UpdateComplaintStatusRequest:
      type: object
      properties:
        statusId: { type: integer, format: int64 }
        note: { type: string, nullable: true }
      required: [statusId]
    AssignComplaintRequest:
      type: object
      properties:
        assignedToUserId: { type: integer, format: int64 }
        note: { type: string, nullable: true }
        dueDate: { type: string, format: date-time, nullable: true }
      required: [assignedToUserId]
    EscalateComplaintRequest:
      type: object
      properties:
        reason: { type: string }
        escalatedToUserId: { type: integer, format: int64 }
        escalationType: { type: string, example: Manual }
      required: [reason, escalatedToUserId, escalationType]
    ResolveComplaintRequest:
      type: object
      properties:
        resolutionSummary: { type: string }
        rootCause: { type: string, nullable: true }
        fixApplied: { type: string, nullable: true }
      required: [resolutionSummary]
...